+++ src/models/FieldInterface.ts
@@ -74,7 +74,7 @@ export interface FieldInterface extends BaseInterface {
   clear(deep?: boolean): void;
   reset(deep?: boolean): void;
   focus(): void;
-  showErrors(show: boolean): void;
+  showErrors(show?: boolean, deep?: boolean): void;
   showAsyncErrors(): void;
   observeValidationOnBlur(): void;
   observeValidationOnChange(): void;